Deleted temporary files at program exit
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@2646 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/tools/OperatingSystem.java
@@ -20,16 +20,12 @@
 package com.eteks.sweethome3d.tools;
 
 import java.io.File;
-import java.io.FileFilter;
 import java.io.IOException;
 import java.util.MissingResourceException;
 import java.util.ResourceBundle;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.UUID;
 
 import com.apple.eio.FileManager;
-import com.eteks.sweethome3d.model.Home;
 
 /**
  * Tools used to test current user operating system.
@@ -111,65 +107,46 @@ public class OperatingSystem {
    * @throws IOException if the file couldn't be created
    */
   public static File createTemporaryFile(String prefix, String suffix) throws IOException {
-    File temporaryFolder = getDefaultTemporaryFolder();
+    File temporaryFolder = getDefaultTemporaryFolder(true);
     File temporaryFile = File.createTempFile(prefix, suffix, temporaryFolder);
     temporaryFile.deleteOnExit();
     return temporaryFile;
   }
-  
+
+  /**
+   * Deletes all the temporary files created with {@link #createTemporaryFile(String, String) createTemporaryFile}.
+   */
+  public static void deleteTemporaryFiles() {
+    try {
+      File temporaryFolder = getDefaultTemporaryFolder(false);
+      if (temporaryFolder != null) {
+        for (File temporaryFile : temporaryFolder.listFiles()) {
+          temporaryFile.delete();
+        }
+        temporaryFolder.delete();
+      }
+    } catch (IOException ex) {
+      // Ignore temporary folder that can't be found
+    }
+  }
+
   /**
    * Returns the default folder used to store temporary files created in the program.
    */
-  private synchronized static File getDefaultTemporaryFolder() throws IOException {
+  private synchronized static File getDefaultTemporaryFolder(boolean create) throws IOException {
     if (TEMPORARY_SUB_FOLDER != null) {
       File temporaryFolder = new File(getDefaultApplicationFolder(), TEMPORARY_SUB_FOLDER);
-      final String versionPrefix = Home.CURRENT_VERSION + "-";
-      final File sessionTemporaryFolder = new File(temporaryFolder, 
-          versionPrefix + TEMPORARY_SESSION_SUB_FOLDER);      
+      final File sessionTemporaryFolder = new File(temporaryFolder, TEMPORARY_SESSION_SUB_FOLDER);      
       if (!sessionTemporaryFolder.exists()) {
-        // Retrieve existing folders working with same Sweet Home 3D version in temporary folder
-        final File [] siblingTemporaryFolders = temporaryFolder.listFiles(new FileFilter() {
-            public boolean accept(File file) {
-              return file.isDirectory() 
-                  && file.getName().startsWith(versionPrefix);
-            }
-          });
-        
-        // Create temporary folder  
-        if (!sessionTemporaryFolder.mkdirs()) {
-          throw new IOException("Can't create temporary folder " + sessionTemporaryFolder);
+        if (create) {
+          // Create temporary folder  
+          if (!sessionTemporaryFolder.mkdirs()) {
+            throw new IOException("Can't create temporary folder " + sessionTemporaryFolder);
+          } else {
+            sessionTemporaryFolder.deleteOnExit();
+          }
         } else {
-          sessionTemporaryFolder.deleteOnExit();
-        }
-        
-        // Launch a timer that updates modification date of the temporary folder each 10 s
-        final long delay = 10000;
-        new Timer(true).schedule(new TimerTask() {
-            @Override
-            public void run() {
-              // Ensure modification date is always growing in case system time was adjusted
-              sessionTemporaryFolder.setLastModified(Math.max(System.currentTimeMillis(),
-                  sessionTemporaryFolder.lastModified() + delay));
-            }
-          }, delay, delay);
-        
-        if (siblingTemporaryFolders != null
-            && siblingTemporaryFolders.length > 0) {
-          // Launch a timer that will delete out dated temporary folders in 20 s
-          new Timer(true).schedule(new TimerTask() {
-              @Override
-              public void run() {
-                long now = System.currentTimeMillis();
-                for (File siblingTemporaryFolder : siblingTemporaryFolders) {
-                  if (now - siblingTemporaryFolder.lastModified() > 2 * delay) {
-                    for (File temporaryFile : siblingTemporaryFolder.listFiles()) {
-                      temporaryFile.delete();
-                    }
-                    siblingTemporaryFolder.delete();
-                  }
-                }
-              }
-            }, 2 * delay);
+          return null;
         }
       }
       return sessionTemporaryFolder;
